+++ src/main/java/com/structurizr/dsl/StructurizrDslParser.java
@@ -732,6 +732,11 @@ public final class StructurizrDslParser extends StructurizrDslTokens {
                             new DocsParser().parse(getContext(ContainerDslContext.class), dslFile, tokens);
                         }
 
+                    } else if (DOCS_TOKEN.equalsIgnoreCase(firstToken) && inContext(ComponentDslContext.class)) {
+                        if (!restricted) {
+                            new DocsParser().parse(getContext(ComponentDslContext.class), dslFile, tokens);
+                        }
+
                     } else if (ADRS_TOKEN.equalsIgnoreCase(firstToken) && inContext(WorkspaceDslContext.class)) {
                         if (!restricted) {
                             new AdrsParser().parse(getContext(WorkspaceDslContext.class), dslFile, tokens);
@@ -745,6 +750,11 @@ public final class StructurizrDslParser extends StructurizrDslTokens {
                     } else if (ADRS_TOKEN.equalsIgnoreCase(firstToken) && inContext(ContainerDslContext.class)) {
                         if (!restricted) {
                             new AdrsParser().parse(getContext(ContainerDslContext.class), dslFile, tokens);
+                        }
+
+                    } else if (ADRS_TOKEN.equalsIgnoreCase(firstToken) && inContext(ComponentDslContext.class)) {
+                        if (!restricted) {
+                            new AdrsParser().parse(getContext(ComponentDslContext.class), dslFile, tokens);
                         }
 
                     } else if (CONSTANT_TOKEN.equalsIgnoreCase(firstToken)) {